% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_performance_metrics.R
\name{clac_MSE}
\alias{clac_MSE}
\title{Compute mean squared error}
\usage{
clac_MSE(y_truth, y_pred)
}
\arguments{
\item{y_truth}{A numeric vector containing the gorund truth}

\item{y_pred}{A numeric vector containing the estimated values}
}
\description{
Compute mean squared error
}
\author{
Priyanga Dilini Talagala
}
