% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_HD.R
\name{NN_HD}
\alias{NN_HD}
\title{NN-HD algorithm}
\usage{
NN_HD(dataset)
}
\arguments{
\item{dataset}{A multivariate dataset containing numerical variables}
}
\description{
This algorithm is inspired by the HDoutliers algorithm
 which is an unsupervised outlier
 detection algorithm that searches for outliers in high dimensional
 data assuming there is a large distance between outliers and the
 typical data.  Nearest neighbor distances between
 points are used to detect outliers. However, variables with large
 variance can bring disproportional influence on Euclidean distance
calculation. Therefore, the columns of the data sets are first
normalized such that the data are bounded by the unit hyper-cube.
The nearest neighbor distances are then calculated for each
observation.  In contrast to the implementation of HDoutliers
algorithm available in the \code{\link[HDoutliers]{HDoutliers}}
package, NN_HD now generates outlier scores instead of labels for
each observation.
}
\references{
Wilkinson, L. (2016). Visualizing outliers. \url{https://www.cs.uic.edu/~wilkinson/Publications/outliers.pdf}
}
\author{
Priyanga Dilini Talagala
}
