% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_series.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Make a matrix of plots with a given data set}
\usage{
plot_pairs(data)
}
\arguments{
\item{data}{A dataframe. "Timestamp" column give the timestamps and "type" column gives
types of the data points (outlier, neighbour, typical).}
}
\value{
A graphical representation of the matrix plot
}
\description{
provide a matrix plot and mark anomalous point in red colour and
neighbouring points in green colour
}
\author{
Priyanga Dilini Talagala
}
