% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_series.R
\name{plot_series}
\alias{plot_series}
\title{Plot Multivariate time series}
\usage{
plot_series(data, title)
}
\arguments{
\item{data}{A dataframe. "Timestamp" column give the timestamps and "type" column gives
types of the data points (outlier, neighbour, typical).}

\item{title}{A character string. This sis the main title of the plot}
}
\value{
A graphical representation of the multivariate time series
}
\description{
Plot multivariate time series and mark anomalous point in red colour and
neighbouring points in green colour
}
\author{
Priyanga Dilini Talagala
}
