% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_series.R
\name{plot_time_gap}
\alias{plot_time_gap}
\title{Plot the time gap}
\usage{
plot_time_gap(data)
}
\arguments{
\item{data}{A dataframe of one or more than two colums."Timestamp" column give the timestamps.}
}
\value{
A graphical representation of the time gaps
}
\description{
Provide graphical representation of time gap. Useful when dealing
with irregular time series
}
\author{
Priyanga Dilini Talagala
}
