% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sandy_creek}
\alias{sandy_creek}
\title{Water Quality Sensor data - Sandy_Creek}
\format{A data frame with 15231 rows and 15 variables:
\describe{
\item{Timestamp }{Time Stamps}
\item{Lmain_Level_m }{Lmain_Level_m }
\item{Lmain_BattVMin}{Lmain_BattVMin}
\item{Lmain_Temp_C}{Lmain_Temp_C}
\item{Lmain_PestFrVMin}{Lmain_PestFrVMin}
\item{Lmain_PumpBattVMin}{Lmain_PumpBattVMin}
\item{Lsonde_Cond_uscm}{Lsonde_Cond_uscm}\
\item{Lsonde_DO_mgL}{Lsonde_DO_mgL}
\item{Lsonde_DO_pSat}{Lsonde_DO_pSat}
\item{Lsonde_fDOM_QSU}{Lsonde_fDOM_QSU}
\item{Lsonde_fDOM_Raw}{Lsonde_fDOM_Raw}
\item{Lsonde_fDOM_RFU}{Lsonde_fDOM_RFU}
\item{Lsonde_Level_m}{Lsonde_Level_m}
\item{Lsonde_Temp_C}{Lsonde_Temp_C}
\item{Lsonde_Turb_NTU}{Lsonde_Turb_NTU}
}}
\usage{
sandy_creek
}
\description{
A multivariate dataset containing the variables obtained using water quality
sensors from Sandy Creek .
}
\keyword{datasets}
