% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_performance_metrics.R
\name{calc_performance_metrics}
\alias{calc_performance_metrics}
\title{Compute performance metrics}
\usage{
calc_performance_metrics(y_truth, y_output, pos_label, neg_label,
  print_out = TRUE)
}
\arguments{
\item{y_truth}{A character vector containing the gorund truth}

\item{y_output}{a character vector containing the predicted labels from the algorithm}

\item{pos_label}{A character string. Label used to indicate the outliers in the original dataframe.}

\item{neg_label}{A character string. Label used to indicate the typical values in the original dataframe.}

\item{print_out}{If TRUE, output will be printed to console.}
}
\value{
A list with the following elements:
\item{TN}{True negatives}
\item{FN}{False negatives}
\item{FP}{False positives}
\item{TP}{True positives}
\item{Accuracy}{Accuracy}
\item{Error_Rate}{Error Rate}
\item{Sensitivity}{Sensitivity}
\item{Specificity}{Specificity}
\item{Precision}{Precision}
\item{Recall}{Recall}
\item{F_Measure}{F Measure}
\item{Optimised_Precision}{Optimised Precision}
}
\description{
Computes various measure to evaluate the performance of an algorithm
}
\examples{
true_labels <- c("out", "out", "normal", "out", "normal", "normal",
                 "normal", "normal", "normal", "normal")
output <- c("out", "normal", "normal", "normal", "out", "out",
            "normal", "normal", "normal", "normal")
out<- calc_performance_metrics(y_truth = true_labels, y_output = output,
                              pos_label = "out", neg_label = "normal")

}
\author{
Priyanga Dilini Talagala
}
