% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Apply different transformations to the original series}
\usage{
transform_data(data, time_bound = 90)
}
\arguments{
\item{data}{A dataframe. This dataframe contains two columns: Timestamp, type in addition
to the variables that need to be transformed}

\item{time_bound}{A positive constant. This is to reduce the effect coming from too
small time gaps when calculating derivatives.}
}
\value{
a dataframe with the original and transformed series
}
\description{
This function apply different transformations to the original variables.
This data preprocessing step was incorporated with the aim of highlighting  different
types of anomalies such as sudden isolated spikes, sudden isolated drops, sudden shifts,
impossible values (negative values) and out of range values etc
}
\author{
Priyanga Dilini Talagala
}
