% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Apply different transformations to the original series}
\usage{
transform_data(data, time_bound = 90, regular = FALSE,
  time_col = "Timestamp")
}
\arguments{
\item{data}{A dataframe. This dataframe contains a seperate column for Timestamp, in addition
to the variables that need to be transformed}

\item{time_bound}{A positive constant. This is to reduce the effect coming from too
small time gaps when calculating derivatives.}

\item{regular}{Regular time interval (TRUE) or irregular (FALSE)}

\item{time_col}{A quoted string to specify the column name of the timestamp}
}
\value{
A tsibble object with the original and the transformed series
}
\description{
This function apply different transformations to the original variables.
This data preprocessing step was incorporated with the aim of highlighting  different
types of anomalies such as sudden isolated spikes, sudden isolated drops, sudden shifts,
impossible values (negative values) and out of range values etc
}
\examples{
data <- data_sandy_anom[,c("Timestamp", "Cond", "Tur", "Level")]
data <- tidyr::drop_na(data)
trans_data <- oddwater::transform_data(data)
}
\author{
Priyanga Dilini Talagala
}
